Simplify PebbleBridge implementation
- Move inline JS shim from PebbleBridgeManager to assets/bridge-shim.js
- Share a single Ktor/OkHttp HttpClient between BridgeFetch and BridgeWebSocket
- Add WebSocket subprotocol support via Sec-WebSocket-Protocol header
- Replace org.json.JSONObject.quote and manual escapeJson with kotlinx.serialization
- Add dispose() path that tears down without firing onClose
- Remove runBlocking(Dispatchers.Main) in Android factory
- Restore localStorage after page finishes loading (LoadingState.Finished)
